Rename style to styles
