+++ src/components/InlineCodeBlock/index.ios.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import {View} from 'react-native';
-import styles, {webViewStyles} from '../../style/StyleSheet';
+import styles, {webViewStyles} from '../../styles/StyleSheet';
 
 const propTypes = {
     children: PropTypes.node.isRequired,
