+++ src/components/InlineCodeBlock/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import {Text} from 'react-native';
-import {webViewStyles} from '../../style/StyleSheet';
+import {webViewStyles} from '../../styles/StyleSheet';
 
 const propTypes = {
     children: PropTypes.node.isRequired,
