+++ src/pages/home/report/ReportActionItemSingle.js
@@ -5,7 +5,7 @@ import _ from 'underscore';
 import ReportActionPropTypes from './ReportActionPropTypes';
 import ReportActionItemMessage from './ReportActionItemMessage';
 import ReportActionItemFragment from './ReportActionItemFragment';
-import styles from '../../../style/StyleSheet';
+import styles from '../../../styles/StyleSheet';
 import CONST from '../../../CONST';
 import ReportActionItemDate from './ReportActionItemDate';
