+++ src/pages/home/sidebar/AppLinks/index.native.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from '../../../../style/StyleSheet';
+import styles from '../../../../styles/StyleSheet';
 import openURLInNewTab from '../../../../libs/openURLInNewTab';
 import Text from '../../../../components/Text';
