+++ src/pages/home/sidebar/ChatSwitcherSearchForm.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {Image, TouchableOpacity, View} from 'react-native';
 import PropTypes from 'prop-types';
-import styles, {colors} from '../../../style/StyleSheet';
+import styles, {colors} from '../../../styles/StyleSheet';
 import logoCircle from '../../../../assets/images/expensify-logo-round.png';
 import TextInputWithFocusStyles from '../../../components/TextInputWithFocusStyles';
 import iconX from '../../../../assets/images/icon-x.png';
