+++ src/pages/home/sidebar/SidebarView.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {View} from 'react-native';
 import PropTypes from 'prop-types';
-import styles from '../../../style/StyleSheet';
+import styles from '../../../styles/StyleSheet';
 import SidebarBottom from './SidebarBottom';
 import SidebarLinks from './SidebarLinks';
 import SafeAreaInsetPropTypes from '../../SafeAreaInsetPropTypes';
